# Conflicts:
#	apps/theming/lib/ThemingDefaults.php

+++ tests/lib/Repair/RepairCollationTest.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * SPDX-FileCopyrightText: 2016-2024 Nextcloud GmbH and Nextcloud contributors
  * SPDX-FileCopyrightText: 2016 ownCloud, Inc.
@@ -12,6 +13,7 @@ use OC\Repair\Collation;
 use OCP\IConfig;
 use OCP\IDBConnection;
 use OCP\Migration\IOutput;
+use OCP\Server;
 use PHPUnit\Framework\MockObject\MockObject;
 use Psr\Log\LoggerInterface;
 use Test\TestCase;
@@ -29,10 +31,9 @@ class TestCollationRepair extends Collation {
 /**
  * Tests for the converting of MySQL tables to InnoDB engine
  *
- * @group DB
- *
  * @see \OC\Repair\RepairMimeTypes
  */
+#[\PHPUnit\Framework\Attributes\Group('DB')]
 class RepairCollationTest extends TestCase {
 
 	private TestCollationRepair $repair;
@@ -45,14 +46,14 @@ class RepairCollationTest extends TestCase {
 	protected function setUp(): void {
 		parent::setUp();
 
-		$this->connection = \OCP\Server::get(ConnectionAdapter::class);
-		$this->config = \OCP\Server::get(IConfig::class);
+		$this->connection = Server::get(ConnectionAdapter::class);
 		if ($this->connection->getDatabaseProvider() !== IDBConnection::PLATFORM_MYSQL) {
 			$this->markTestSkipped('Test only relevant on MySql');
+			return;
 		}
 
 		$this->logger = $this->createMock(LoggerInterface::class);
-
+		$this->config = Server::get(IConfig::class);
 		$dbPrefix = $this->config->getSystemValueString('dbtableprefix');
 		$this->tableName = $this->getUniqueID($dbPrefix . '_collation_test');
 		$this->connection->prepare("CREATE TABLE $this->tableName(text VARCHAR(16)) COLLATE utf8_unicode_ci")->execute();
@@ -61,7 +62,11 @@ class RepairCollationTest extends TestCase {
 	}
 
 	protected function tearDown(): void {
-		$this->connection->getInner()->createSchemaManager()->dropTable($this->tableName);
+		$this->connection = Server::get(ConnectionAdapter::class);
+		if ($this->connection->getDatabaseProvider() === IDBConnection::PLATFORM_MYSQL) {
+			// tear down only needed on MySQL
+			$this->connection->getInner()->createSchemaManager()->dropTable($this->tableName);
+		}
 		parent::tearDown();
 	}
